NEW: add http server in BambuStudio
(cherry picked from commit 9ee36044158856e433b19e407094120bd15f4aa0)

+++ src/imgui/imgui_internal.h
@@ -391,7 +391,7 @@ IM_MSVC_RUNTIME_CHECKS_OFF
 static inline float  ImPow(float x, float y)    { return powf(x, y); }          // DragBehaviorT/SliderBehaviorT uses ImPow with either float/double and need the precision
 static inline double ImPow(double x, double y)  { return pow(x, y); }
 static inline float  ImLog(float x)             { return logf(x); }             // DragBehaviorT/SliderBehaviorT uses ImLog with either float/double and need the precision
-static inline double ImLog(double x)            { return log(x); }
+static inline double ImLog(double x)            { return logf(x); }
 static inline int    ImAbs(int x)               { return x < 0 ? -x : x; }
 static inline float  ImAbs(float x)             { return fabsf(x); }
 static inline double ImAbs(double x)            { return fabs(x); }
